Use `Serialize(nBytes)` instead of doing it by hand

+++ Sts1CobcSw/Periphery/Rf.hpp
@@ -29,8 +29,8 @@ auto Initialize(TxType txType) -> void;
 auto ReadPartNumber() -> std::uint16_t;
 
 auto SetTxType(TxType txType) -> void;
-// TODO: Return type should be void;
-auto Send(void const * data, std::size_t nBytes) -> bool;
+// TODO: Maybe return a Result<void, E> instead
+auto Send(void const * data, std::uint16_t nBytes) -> bool;
 // TODO: Replace this by Receive(void * data, std::size_t nBytes) -> void;
 auto ReceiveTestData() -> std::array<Byte, maxRxBytes>;
 }
